Refactor the auth.controllers.ts
- significantly reduce total lines of code in the auth.controllers.ts
  file by moving several more genarlized controllers to a new file named
user.controllers.ts
- this was because i needed to make the file leaner and easier to parse
  through. all the major controllers associated with authentication are
still containe within, such as registering a new user, loggin in an
existing user, changing user password e.t.c
- the controllers moved include:
	*updateUser 	*getUser	* getUsers
	*getUserById	* getUserByEmailOrUserName
	*getUserByRole
- these seemed to focus less on the authentication bit and more on the
  user.
