+++ collections/auth.collections.http
@@ -42,49 +42,8 @@ Content-Type: application/json
     "confirmNewPassword": "Shanara17!"
 }
 
-######################
-# UPDATE USER DETAILS
-PATCH http://localhost:4000/auth/update/917b6e63-3ed4-4b42-9d67-e16216039c74
-Content-Type: application/json
-
-{
-    "firstName": "Minimus",
-    "lastName": "Lancelot",
-    "userName": "mini",
-    "email": "lance@gmail.com",
-    "contactNumber": "073-550-1059"
-}
 
 ######################
 # DELETE USER
 DELETE http://localhost:4000/auth/delete/917b6e63-3ed4-4b42-9d67-e16216039c74
 
-
-######################
-# GET USER BY ID
-GET http://localhost:4000/auth/search/b3b4500b-fc78-4237-9696-cf67ba22d8d0
-
-
-######################
-# GET USERS
-GET http://localhost:4000/auth/
-
-
-######################
-# GET USERS BY ROLE 
-GET http://localhost:4000/auth/search/role/
-Content-Type: application/json
-
-{
-    "role":"tenant"
-}
-
-
-######################
-# GET USERS BY EMAIL OR USERNAME 
-GET http://localhost:4000/auth/search/mail-or-name
-Content-Type: application/json
-
-{
-    "emailOrUserName": "bri@gmail.com"
-}
