+++ src/server.ts
@@ -1,6 +1,7 @@
 import express,{ json }  from "express"
 import dotenv from "dotenv"
 import authRouter from "./api-v1/routes/auth.routes"
+import userRouter from "./api-v1/routes/user.routes"
 
 // configure path to dotenv variables
 dotenv.config()
@@ -13,6 +14,7 @@ app.use(json()) // add a body to requests
 
 // define URLs within the application
 app.use('/auth', authRouter)
+app.use('/user', userRouter)
